Remove separate Java-packages for replenishment events and commands.
