+++ model/src/main/proto/spine_examples/shareaware/wallet/replenishment_commands.proto
@@ -31,14 +31,13 @@ package spine_examples.shareaware.wallet;
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.shareaware.spine.io";
-option java_package = "io.spine.examples.shareaware.wallet.replenishmentcommand";
-option java_outer_classname = "CommandsProto";
+option java_package = "io.spine.examples.shareaware.wallet.command";
+option java_outer_classname = "ReplenishmentCommandsProto";
 option java_multiple_files = true;
 
 import "spine_examples/shareaware/identifiers.proto";
 import "spine/money/money.proto";
 import "spine_examples/shareaware/wallet/iban.proto";
-import "spine/core/user_id.proto";
 
 // A command to replenish the wallet.
 message ReplenishWallet {
