+++ model/src/main/proto/spine_examples/shareaware/wallet/replenishment_events.proto
@@ -31,8 +31,8 @@ package spine_examples.shareaware.wallet;
 import "spine/options.proto";
 
 option (type_url_prefix) = "type.shareaware.spine.io";
-option java_package = "io.spine.examples.shareaware.wallet.replenishmentevent";
-option java_outer_classname = "EventsProto";
+option java_package = "io.spine.examples.shareaware.wallet.event";
+option java_outer_classname = "ReplenishmentEventsProto";
 option java_multiple_files = true;
 
 import "spine/money/money.proto";
