+++ server/src/test/java/io/spine/examples/shareaware/server/given/WalletTestEnv.java
@@ -4,7 +4,7 @@ import io.spine.examples.shareaware.ReplenishmentId;
 import io.spine.examples.shareaware.WalletId;
 import io.spine.examples.shareaware.wallet.Iban;
 import io.spine.examples.shareaware.wallet.command.CreateWallet;
-import io.spine.examples.shareaware.wallet.replenishmentcommand.ReplenishWallet;
+import io.spine.examples.shareaware.wallet.command.ReplenishWallet;
 import io.spine.money.Money;
 import io.spine.testing.core.given.GivenUserId;
 import io.spine.testing.server.blackbox.BlackBoxContext;
