+++ server/src/test/java/io/spine/examples/shareaware/server/wallet/WalletReplenishmentTest.java
@@ -38,9 +38,9 @@ import io.spine.examples.shareaware.wallet.WalletBalance;
 import io.spine.examples.shareaware.wallet.WalletReplenishment;
 import io.spine.examples.shareaware.wallet.command.RechargeBalance;
 import io.spine.examples.shareaware.wallet.event.BalanceRecharged;
-import io.spine.examples.shareaware.wallet.replenishmentcommand.ReplenishWallet;
-import io.spine.examples.shareaware.wallet.replenishmentevent.WalletNotReplenished;
-import io.spine.examples.shareaware.wallet.replenishmentevent.WalletReplenished;
+import io.spine.examples.shareaware.wallet.command.ReplenishWallet;
+import io.spine.examples.shareaware.wallet.event.WalletNotReplenished;
+import io.spine.examples.shareaware.wallet.event.WalletReplenished;
 import io.spine.money.Currency;
 import io.spine.money.Money;
 import io.spine.server.BoundedContextBuilder;
